Add ParserTest file
